Start refactor of the user security model
